move keypress code into it's own file under app
