+++ src/components/Main.vue
@@ -66,22 +66,12 @@
 
 <script>
 import capitalize from 'lodash/capitalize';
-import {
-  mapMutations as _mapMutations,
-  createNamespacedHelpers,
-  mapState as _mapState,
-  mapGetters as _mapGetters,
-  mapActions
-} from 'vuex';
-const { mapState, mapGetters, mapMutations } = createNamespacedHelpers(
-  'keymap'
-);
+import { mapMutations, mapState, mapGetters, mapActions } from 'vuex';
 import ControllerTop from '@/components/ControllerTop';
 import StatusPanel from '@/components/StatusPanel';
 import ControllerBottom from '@/components/ControllerBottom';
 import VisualKeymap from '@/components/VisualKeymap';
 import LayerControl from '@/components/LayerControl';
-import * as jquery from '@/jquery';
 
 export default {
   name: 'Main',
@@ -94,10 +84,10 @@ export default {
     LayerControl
   },
   computed: {
-    ..._mapState('app', ['appInitialized', 'configuratorSettings']),
-    ..._mapGetters('app', ['keyCount']),
-    ...mapState(['continuousInput']),
-    ...mapGetters(['colorwayIndex', 'colorways', 'size']),
+    ...mapState('app', ['appInitialized', 'configuratorSettings']),
+    ...mapGetters('app', ['keyCount']),
+    ...mapState('keymap', ['continuousInput']),
+    ...mapGetters('keymap', ['colorwayIndex', 'colorways', 'size']),
     curIndex: {
       get() {
         return this.colorwayIndex;
@@ -136,9 +126,9 @@ export default {
     }
   },
   methods: {
-    ...mapActions('app', ['setFavoriteColor']),
-    ...mapMutations(['nextColorway']),
-    ..._mapMutations('app', ['resetListener']),
+    ...mapActions('app', ['setFavoriteColor', 'initKeypressListener']),
+    ...mapMutations('keymap', ['nextColorway']),
+    ...mapMutations('app', ['resetListener']),
     favColor() {
       if (this.isFavoriteColor) {
         this.setFavoriteColor('');
@@ -147,8 +137,8 @@ export default {
       }
     }
   },
-  mounted() {
-    jquery.init();
+  async mounted() {
+    await this.initKeypressListener();
     // Loading favorite color
     if (this.configuratorSettings.favoriteColor) {
       const favoriteColor = this.configuratorSettings.favoriteColor.toLowerCase();
